backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "google-generativeai>=0.7.2",
    "openai-agents>=0.6.3",
    "qdrant-client>=1.16.2",
    "langchain-google-genai>=2.0.0",
    "langchain-text-splitters>=0.3.0",
    "pytest>=8.0.0",
    "requests>=2.31.0",
    "asyncpg>=0.31.0",
    "sqlmodel>=0.0.27",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "python-multipart>=0.0.21",
]
